split fixtures into folder
