+++ tests/fixtures/__init__.py
@@ -1,0 +1,3 @@
+from .paths import *
+from .generation import *
+from .models import *
